+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Image Modification Changelog
 
-## [Bug Fixes & New Filters] - TBD
+## [Remove Background] - {PR_MERGE_DATE}
+
+- Added 'Remove Background' command to remove the background from selected images
+
+## [Bug Fixes & New Filters] - 2025-01-30
 
 - Improved memory management when previewing filters, reducing the likelihood of out-of-memory errors
 - Added setting to disable live filter previews